fix typo; move scripts to cleanup namespace; fix config paths when installed as package
