+++ acegen/scripts/a2c/a2c.py
@@ -3,8 +3,8 @@ import json
 import os
 import random
 import shutil
-from copy import deepcopy
 from pathlib import Path
+from importlib import resources
 
 import hydra
 import numpy as np
@@ -17,16 +17,14 @@ from acegen.rl_env import generate_complete_smiles, SMILESEnv
 from acegen.scoring_functions import custom_scoring_functions, Task
 from acegen.vocabulary import SMILESVocabulary
 from omegaconf import OmegaConf
-from tensordict.utils import isin
+from tensordict import TensorDict
+from torch.distributions.kl import kl_divergence
+from torchrl.data import LazyTensorStorage, TensorDictReplayBuffer
+from torchrl.data.replay_buffers.samplers import SamplerWithoutReplacement
 
-from torchrl.data import (
-    LazyTensorStorage,
-    PrioritizedSampler,
-    TensorDictMaxValueWriter,
-    TensorDictPrioritizedReplayBuffer,
-    TensorDictReplayBuffer,
-)
 from torchrl.envs import InitTracker, TensorDictPrimer, TransformedEnv
+from torchrl.objectives import A2CLoss
+from torchrl.objectives.value.advantages import GAE
 from torchrl.record.loggers import get_logger
 
 try:
@@ -40,7 +38,7 @@ except ImportError as err:
     MOLSCORE_ERR = err
 
 
-@hydra.main(config_path=".", config_name="config_denovo", version_base="1.2")
+@hydra.main(config_path=str(resources.files("acegen.scripts.a2c")), config_name="config_denovo", version_base="1.2")
 def main(cfg: "DictConfig"):
 
     # Set seeds
@@ -77,7 +75,7 @@ def main(cfg: "DictConfig"):
                 include=cfg.molscore_include,
             )
             for task in MSB:
-                run_reinvent(cfg, task)
+                run_a2c(cfg, task)
         else:
             # Save molscore output. Also redirect output to save_dir
             cfg.molscore = shutil.copy(cfg.molscore, save_dir)
@@ -89,15 +87,15 @@ def main(cfg: "DictConfig"):
                 budget=cfg.total_smiles,
                 output_dir=os.path.abspath(save_dir),
             )
-            run_reinvent(cfg, task)
+            run_a2c(cfg, task)
     elif cfg.get("custom_task", None):
         task = Task(custom_scoring_functions[cfg.custom_task], budget=cfg.total_smiles)
-        run_reinvent(cfg, task)
+        run_a2c(cfg, task)
     else:
         raise ValueError("No scoring function specified.")
 
 
-def run_reinvent(cfg, task):
+def run_a2c(cfg, task):
 
     # Get available device
     device = (
@@ -106,7 +104,14 @@ def run_reinvent(cfg, task):
 
     # Get model and vocabulary checkpoints
     if cfg.model in models:
-        create_actor, _, _, voc_path, ckpt_path, tokenizer = models[cfg.model]
+        (
+            create_actor,
+            create_critic,
+            create_shared,
+            voc_path,
+            ckpt_path,
+            tokenizer,
+        ) = models[cfg.model]
     else:
         raise ValueError(f"Unknown model type: {cfg.model}")
 
@@ -123,27 +128,50 @@ def run_reinvent(cfg, task):
     # Create models
     ####################################################################################################################
 
-    ckpt = torch.load(ckpt_path)
+    # Create actor and critic networks
+    if cfg.shared_nets:
+        (
+            actor_training,
+            actor_inference,
+            critic_training,
+            critic_inference,
+        ) = create_shared(vocabulary_size=len(vocabulary))
+    else:
+        actor_training, actor_inference = create_actor(len(vocabulary))
+        critic_training, critic_inference = create_critic(len(vocabulary))
 
-    actor_training, actor_inference = create_actor(vocabulary_size=len(vocabulary))
+    # Load pretrained weights
+    ckpt = torch.load(ckpt_path)
     actor_inference.load_state_dict(
         adapt_state_dict(ckpt, actor_inference.state_dict())
     )
     actor_training.load_state_dict(adapt_state_dict(ckpt, actor_training.state_dict()))
     actor_inference = actor_inference.to(device)
     actor_training = actor_training.to(device)
+    critic_training = critic_training.to(device)
 
-    prior = deepcopy(actor_training)
+    # Define prior
+    prior, _ = create_actor(len(vocabulary))
+    prior = prior.to(device)
+    prior.load_state_dict(adapt_state_dict(ckpt, prior.state_dict()))
 
     # Create RL environment
     ####################################################################################################################
 
-    # For RNNs, create a transform to populate initial tensordict with recurrent states equal to 0.0
     rhs_primers = []
-    if hasattr(actor_training, "rnn_spec"):
+    # if rnn's, create a transform to populate initial tensordict with recurrent states equal to 0.0
+    if cfg.shared_nets and hasattr(actor_training, "rnn_spec"):
         primers = actor_training.rnn_spec.expand(cfg.num_envs)
-        rhs_primers.append(TensorDictPrimer(primers))
+        rhs_primers = [TensorDictPrimer(primers)]
+    elif hasattr(actor_training, "rnn_spec"):
+        actor_primers = actor_training.rnn_spec.expand(cfg.num_envs)
+        critic_primers = critic_training.rnn_spec.expand(cfg.num_envs)
+        rhs_primers = [
+            TensorDictPrimer(actor_primers),
+            TensorDictPrimer(critic_primers),
+        ]
 
+    # Define environment kwargs
     env_kwargs = {
         "start_token": vocabulary.start_token_index,
         "end_token": vocabulary.end_token_index,
@@ -152,6 +180,7 @@ def run_reinvent(cfg, task):
         "device": device,
     }
 
+    # Define environment creation function
     def create_env_fn():
         """Create a single RL rl_env."""
         env = SMILESEnv(**env_kwargs)
@@ -163,23 +192,39 @@ def run_reinvent(cfg, task):
 
     env = create_env_fn()
 
-    # Create replay buffer
+    # Create loss module
     ####################################################################################################################
 
-    storage = LazyTensorStorage(cfg.replay_buffer_size, device=device)
-    experience_replay_buffer = TensorDictReplayBuffer(
-        storage=storage,
-        sampler=PrioritizedSampler(storage.max_size, alpha=1.0, beta=1.0),
-        batch_size=cfg.replay_batch_size,
-        writer=TensorDictMaxValueWriter(rank_key="priority"),
-        priority_key="priority",
+    adv_module = GAE(
+        gamma=cfg.gamma,
+        lmbda=cfg.lmbda,
+        value_network=critic_training,
+        average_gae=True,
+        shifted=True,
+    )
+    loss_module = A2CLoss(
+        actor_network=actor_training,
+        critic_network=critic_training,
+        critic_coef=cfg.critic_coef,
+        entropy_coef=cfg.entropy_coef,
+        loss_critic_type="l2",
+    )
+
+    # Create data storage
+    ####################################################################################################################
+
+    buffer = TensorDictReplayBuffer(
+        storage=LazyTensorStorage(cfg.num_envs, device=device),
+        sampler=SamplerWithoutReplacement(),
+        batch_size=cfg.mini_batch_size,
+        prefetch=4,
     )
 
     # Create optimizer
     ####################################################################################################################
 
     optim = torch.optim.Adam(
-        actor_training.parameters(),
+        loss_module.parameters(),
         lr=cfg.lr,
         eps=cfg.eps,
         weight_decay=cfg.weight_decay,
@@ -190,16 +235,10 @@ def run_reinvent(cfg, task):
 
     logger = None
     if cfg.logger_backend:
-        experiment_name = f"{cfg.agent_name}"
-        try:
-            experiment_name += f"_{task.configs.get('task')}"
-        except AttributeError:
-            experiment_name += "_custom_task"
-
         logger = get_logger(
             cfg.logger_backend,
-            logger_name="reinvent",
-            experiment_name=experiment_name,
+            logger_name="a2c",
+            experiment_name=f"{cfg.agent_name}_{task.configs.get('task')}",
             wandb_kwargs={
                 "config": dict(cfg),
                 "project": cfg.experiment_name,
@@ -212,8 +251,12 @@ def run_reinvent(cfg, task):
     ####################################################################################################################
 
     total_done = 0
-    sigma = cfg.sigma
+    num_updates = 0
+    kl_coef = cfg.kl_coef
+    max_grad_norm = cfg.max_grad_norm
     pbar = tqdm.tqdm(total=cfg.total_smiles)
+    num_mini_batches = cfg.num_envs // cfg.mini_batch_size
+    losses = TensorDict({}, batch_size=[num_mini_batches])
 
     while not task.finished:
 
@@ -233,13 +276,14 @@ def run_reinvent(cfg, task):
             remove_duplicates=True,
         )
 
-        log_info = {}
-        total_done += cfg.num_envs
+        # Update progress bar
         data_next = data.get("next")
         done = data_next.get("done").squeeze(-1)
         pbar.update(done.sum().item())
 
-        # Save info about smiles lengths and rewards
+        # Register smiles lengths and real rewards and total generated smiles
+        log_info = {}
+        total_done += cfg.num_envs
         episode_rewards = data_next["reward"][done]
         episode_length = (data_next["observation"] != 0.0).float().sum(-1).mean()
         if len(episode_rewards) > 0:
@@ -253,87 +297,57 @@ def run_reinvent(cfg, task):
                 }
             )
 
-        data, loss, agent_likelihood = compute_loss(data, actor_training, prior, sigma)
+        # Compute advantage
+        with torch.no_grad():
+            data = adv_module(data)
 
-        # Compute experience replay loss
-        if (
-            cfg.experience_replay
-            and len(experience_replay_buffer) > cfg.replay_batch_size
-        ):
-            replay_batch = experience_replay_buffer.sample()
-            _, replay_loss, replay_agent_likelihood = compute_loss(
-                replay_batch, actor_training, prior, sigma
+        # Add extended_data to buffer
+        buffer.extend(data)
+
+        for j, batch in enumerate(buffer):
+
+            batch = batch.to(device, non_blocking=True)
+
+            # Compute loss
+            mask = batch.get("mask").squeeze(-1)
+            loss = loss_module(batch)
+            loss = loss.named_apply(
+                lambda name, value: (
+                    (value * mask).mean() if name.startswith("loss_") else value
+                ),
+                batch_size=[],
             )
-            loss = torch.cat((loss, replay_loss), 0)
-            agent_likelihood = torch.cat((agent_likelihood, replay_agent_likelihood), 0)
+            loss_sum = (
+                loss["loss_critic"] + loss["loss_objective"] + loss["loss_entropy"]
+            )
+            losses[j] = loss.select(
+                "loss_critic", "loss_entropy", "loss_objective"
+            ).detach()
 
-        # Average loss over the batch
-        loss = loss.mean()
+            # Add KL loss term
+            with torch.no_grad():
+                prior_dist = prior.get_dist(batch)
+                kl_div = kl_divergence(actor_training.get_dist(batch), prior_dist)
+            kl_div = (kl_div * mask.squeeze()).sum(-1).mean(-1)
+            loss_sum += kl_div * kl_coef
+            losses[j] = TensorDict({"kl_div": kl_div.detach().item()}, batch_size=[])
 
-        # Add regularizer that penalizes high likelihood for the entire sequence
-        loss_p = -(1 / agent_likelihood).mean()
-        loss += 5 * 1e3 * loss_p
+            # Update policy
+            loss_sum.backward()
+            torch.nn.utils.clip_grad_norm_(
+                loss_module.parameters(), max_norm=max_grad_norm
+            )
+            optim.step()
+            optim.zero_grad()
+            num_updates += 1
 
-        # Calculate gradients and make an update to the network weights
-        optim.zero_grad()
-        loss.backward()
-        optim.step()
+        losses_mean = losses.apply(lambda x: x.float().mean(), batch_size=[])
+        for key, value in losses_mean.items():
+            log_info.update({f"train/{key}": value.item()})
 
-        # Then add new experiences to the replay buffer
-        if cfg.experience_replay is True:
-
-            replay_data = data.clone()
-
-            # MaxValueWriter is not compatible with storages of more than one dimension.
-            replay_data.batch_size = [replay_data.batch_size[0]]
-
-            # Remove SMILES that are already in the replay buffer
-            if len(experience_replay_buffer) > 0:
-                is_duplicated = isin(
-                    input=replay_data,
-                    key="action",
-                    reference=experience_replay_buffer[:],
-                )
-                replay_data = replay_data[~is_duplicated]
-
-            # Add data to the replay buffer
-            reward = replay_data.get(("next", "reward"))
-            replay_data.set("priority", reward)
-            experience_replay_buffer.extend(replay_data)
-
-        # Log info
         if logger:
             for key, value in log_info.items():
                 logger.log_scalar(key, value, step=total_done)
-
-
-def get_log_prob(data, model):
-    actions = data.get("action")
-    model_in = data.select(*model.in_keys, strict=False)
-    log_prob = model.get_dist(model_in).log_prob(actions)
-    return log_prob
-
-
-def compute_loss(data, model, prior, sigma):
-
-    mask = data.get("mask").squeeze(-1)
-
-    if "prior_log_prob" not in data.keys():
-        with torch.no_grad():
-            prior_log_prob = get_log_prob(data, prior)
-            data.set("prior_log_prob", prior_log_prob)
-    else:
-        prior_log_prob = data.get("prior_log_prob")
-
-    agent_log_prob = get_log_prob(data, model)
-    agent_likelihood = (agent_log_prob * mask).sum(-1)
-    prior_likelihood = (prior_log_prob * mask).sum(-1)
-    score = data.get(("next", "reward")).squeeze(-1).sum(-1)
-
-    augmented_likelihood = prior_likelihood + sigma * score
-    loss = torch.pow((augmented_likelihood - agent_likelihood), 2)
-
-    return data, loss, agent_likelihood
 
 
 if __name__ == "__main__":
