+++ acegen/scripts/pretrain/pretrain.sh
